upgraded astro
