+++ .astro/content-assets.mjs
@@ -1,0 +1,1 @@
+export default new Map();
