+++ .astro/types.d.ts
@@ -1,2 +1,2 @@
 /// <reference types="astro/client" />
-/// <reference path="astro/content.d.ts" />
+/// <reference path="content.d.ts" />
