+++ package.json
@@ -13,17 +13,13 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@astrojs/mdx": "^3.1.7",
-    "@astrojs/react": "^3.6.2",
-    "@astrojs/sitemap": "^3.1.6",
-    "@astrolib/seo": "^1.0.0-beta.6",
-    "@types/react": "^18.3.10",
-    "@types/react-dom": "^18.3.0",
+    "@astrojs/react": "latest",
+    "@astrojs/sitemap": "latest",
+    "@sentry/astro": "^8.51.0",
+    "@spotlightjs/astro": "^2.4.2",
     "animate.css": "^4.1.1",
-    "astro": "^4.15.9",
+    "astro": "^5.1.10",
     "bootstrap": "^5.3.3",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
     "typescript": "^5.6.2"
   }
 }
