+++ src/components/filteredList.astro
@@ -1,5 +1,5 @@
 ---
-
+const { items } = Astro.props;
 ---
 
 <!-- Control buttons -->
