+++ src/layouts/BaseLayout.astro
@@ -1,6 +1,5 @@
 ---
 import Head from "../components/head.astro";
-import Navbar from "../components/navbar.astro";
 import Footer from "../components/footer.astro";
 import Header from "../components/base/header.astro";
 const { title } = Astro.props;
@@ -8,9 +7,12 @@ const { title } = Astro.props;
 
 <!doctype html>
 <html lang="en">
-    <Head />
+    <Head title={title} />
     <body>
         <Header />
+
+        <slot />
+
         <Footer />
     </body>
 </html>
