+++ src/layouts/BlogPost.astro
@@ -1,0 +1,40 @@
+---
+import BaseLayout from "./BaseLayout.astro";
+const { frontmatter } = Astro.props;
+---
+
+<BaseLayout title={frontmatter.title}>
+    <section
+        class="masthead"
+        style="background-image: url('assets/img/post-bg.jpg')"
+    >
+        <div class="container position-relative px-4 px-lg-5">
+            <div class="row gx-4 gx-lg-5 justify-content-center">
+                <div class="col-md-10 col-lg-8 col-xl-7">
+                    <div class="post-heading">
+                        <h1>
+                            {frontmatter.title}
+                        </h1>
+                        <h2 class="subheading">
+                            Problems look mighty small from 150 miles up
+                        </h2>
+                        <span class="meta">
+                            Posted by
+                            <a href="#!">Start Bootstrap</a>
+                            on August 24, 2023
+                        </span>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
+
+    <!-- Post Content-->
+    <article class="mb-4">
+        <div class="container px-4 px-lg-5">
+            <div class="row gx-4 gx-lg-5 justify-content-center">
+                <slot />
+            </div>
+        </div>
+    </article>
+</BaseLayout>
