+++ src/pages/index.astro
@@ -20,12 +20,11 @@ const { animate_css = "../../node_modules/animate.css/animate.min.css" } =
       <div class="row vh-100 justify-content-center">
         <div class="col-11 col-md-8 align-content-center">
           <div class="animate__animated animate__fadeInDown">
-            <h1 class="display-5">Hi,<br>
-            <span class="display-1">I'm Kojo Acquah</span></h1>
+            <h5 class="display-5">Hi,</h5>
+            <h1 class="display-1">I'm Kojo Acquah</h1>
             <h5 class="display-5">
-              Business IT Consultant,<br /> 
-              IT Generalist,<br />
-              based in Nova Scotia, Canada
+              Business IT Consultant & <br />IT Generalist, based in Nova
+              Scotia, Canada
             </h5>
             <p class="fs-3">
               I maximize business and process efficiency through technology
