+++ src/portfolio/test_item.md
@@ -1,0 +1,8 @@
+---
+layout: ../../layouts/BlogPost.astro
+title: Test Item One
+slug: /portfolio/test_item_1
+description: Built a chrome extension to detect a website's underlying tech using the WhatCMS.org API
+---
+
+This is some content for the first post
