+++ src/portfolio/what_cms.md
@@ -1,7 +1,8 @@
 ---
-title: What CMS Is It? Chrome Extensions
+layout: ../../layouts/BlogPost.astro
+title: What CMS Is It Part 1
 slug: /portfolio/what_cms_is_it_extension
 description: Built a chrome extension to detect a website's underlying tech using the WhatCMS.org API
 ---
 
-This is some content
+This is some content for the first post
